+++ src/browser/js/settings.js
@@ -16,9 +16,9 @@ const settings = () => {
     const localize = () => {
         document.title = chrome.i18n.getMessage("settings_pageTitle");
         document.querySelectorAll('[data-locale]').forEach(el => {
-            let text = chrome.i18n.getMessage(el.dataset.locale);
-            if ((typeof text !== 'undefined') && text !== '') {
-                el.innerText = text;
+            let html = chrome.i18n.getMessage(el.dataset.locale);
+            if ((typeof html !== 'undefined') && html !== '') {
+                el.innerHTML = html;
             }
         });
     }